requirements.txt
aniso8601==9.0.1
attrs==22.2.0
click==8.1.3
colorama==0.4.6
Flask==2.1.3
Flask-Cors==3.0.10
flask-restx==0.5.1
Flask-SQLAlchemy==2.5.1
greenlet==2.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
MarkupSafe==2.1.2
marshmallow==3.16.0
packaging==23.0
PyJWT==2.4.0
pyrsistent==0.19.3
python-dotenv==1.0.0
pytz==2022.7.1
six==1.16.0
SQLAlchemy==1.4.0
typing_extensions==4.5.0
Werkzeug==2.1.2
